requirements.txt
absl-py==0.9.0
apex==0.1
astor==0.8.1
atomicwrites==1.4.0
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cx-Freeze==6.1
cxxfilt==0.2.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
freetype-py==2.1.0.post1
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.10
imageio==2.8.0
importlib-metadata==1.6.0
inflect==4.1.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
more-itertools==8.3.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
num2words==0.5.10
numba==0.48.0
numpy==1.18.4
opt-einsum==3.2.1
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
pipreqs==0.4.10
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.0
py==1.8.1
pycparser==2.20
pydub==0.24.0
Pygments==2.6.1
pyparsing==2.4.7
pypng==0.0.20
PyQt5==5.14.2
PyQt5-sip==12.7.2
PyQt5-stubs==5.14.2.2
pyrsistent==0.16.0
pytest==5.4.2
python-dateutil==2.8.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.24.0
resampy==0.2.2
scikit-image==0.17.2
scikit-learn==0.23.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
SoundFile==0.10.3.post1
tensorboard==1.15.0
tensorboardX==1.1
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
tifffile==2020.5.30
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
unicode==2.7
Unidecode==1.1.1
urllib3==1.25.11
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarg==0.1.9
zipp==3.1.0
